#8. Error Handling

+++ db/db_user.py
@@ -2,6 +2,7 @@ from sqlalchemy.orm.session import Session
 from db.hash import Hash
 from schemas import UserBase
 from db.models import DbUser
+from fastapi import HTTPException, status
 
 
 def create_user(db: Session, request: UserBase):
@@ -22,13 +23,23 @@ def get_all_users(db: Session):
 
 
 def get_user(db: Session, id: int):
+    user = db.query(DbUser).filter(DbUser.id == id).first()
+
     # Handle any exception
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail=f'Article with id {id} not found')
+
     return db.query(DbUser).filter(DbUser.id == id).first()
 
 
 def update_user(db: Session, id: int, request: UserBase):
     user = db.query(DbUser).filter(DbUser.id == id)
     # Handle any exception
+    if not user.first():
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail=f'Article with id {id} not found')
+
     user.update({
         DbUser.username: request.username,
         DbUser.email: request.email,
@@ -41,6 +52,10 @@ def update_user(db: Session, id: int, request: UserBase):
 def delete_user(db: Session, id: int):
     user = db.query(DbUser).filter(DbUser.id == id).first()
     # Handle any exception
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail=f'Article with id {id} not found')
+
     db.delete(user)
     db.commit()
